Finish Exercise 4.23

+++ controllers/blogs.js
@@ -1,8 +1,13 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const userExtractor = require("../utils/middleware").userExtractor;
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+  });
   response.json(blogs);
 });
 
@@ -16,20 +21,28 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 });
 
-blogsRouter.post("/", async (request, response) => {
+//  register a middleware only for a specific operation
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const body = request.body;
+
+  const user = await User.findById(request.user);
 
   const newBlog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
+    user: user.id,
   });
 
-  const blog = await newBlog.save();
-  response.status(201).json(blog);
+  const savedBlog = await newBlog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
 });
 
+//  register a middleware only for a specific operation
 blogsRouter.put("/:id", async (request, response) => {
   const body = request.body;
 
@@ -42,13 +55,22 @@ blogsRouter.put("/:id", async (request, response) => {
     runValidators: true,
     context: "query",
   });
-  console.log(updatedBlog);
+
   response.status(204).end();
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+//  register a middleware only for a specific operation
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  blogToDelete = await Blog.findById(request.params.id);
+
+  if (blogToDelete.user.toString() === request.user) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response
+      .status(400)
+      .send({ error: "only the creator can delete his own blog" });
+  }
 });
 
 module.exports = blogsRouter;
